+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerFilterDates.test.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import moment from 'moment';
+import classNames from 'classnames/bind';
 import DatePicker from '../../../DatePicker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 const isWeekday = (date) => {
   const momentDate = moment(date);
@@ -14,10 +18,12 @@ const isWeekday = (date) => {
 };
 
 const DatePickerFilterDates = () => (
-  <DatePicker
-    name="date-input"
-    filterDate={isWeekday}
-  />
+  <div className={cx('content-wrapper')}>
+    <DatePicker
+      name="date-input"
+      filterDate={isWeekday}
+    />
+  </div>
 );
 
 export default DatePickerFilterDates;