+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -7,6 +7,17 @@ Unreleased
 * Changed 12-hour format appearance for mobile and desktop.
   * Meridiem select on Desktop has been changed to buttons which means that the WebdriverIO reference screenshots (if any) and/or any code in the consumer application that accesses the old meridiem select functionality on the Desktop view need to be updated.
 
+4.16.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
+* Added `isInvalidMeridiem` prop 
+* Added `required` prop 
+
 4.15.0 - (October 3, 2019)
 ------------------
 ### Changed