+++ packages/terra-form-validation/CHANGELOG.md
@@ -4,6 +4,12 @@ ChangeLog
 Unreleased
 ----------
 
+1.12.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+* Updated example to consume new `isIncomplete` prop in `terra-form-input`
+
 1.11.0 - (October 3, 2019)
 ------------------
 ### Changed